requirements.txt
fastapi
uvicorn
sqlalchemy
pgvector
pymupdf
openai
celery
redis
PyJWT
pydantic-settings
psycopg2-binary
python-multipart
